+++ typescript/sdk/src/providers/builders/types.ts
@@ -1,11 +1,11 @@
 import type { ChainMetadata } from '../../metadata/chainMetadataTypes.js';
 import type { TypedProvider } from '../ProviderType.js';
-import type { ProviderRetryOptions } from '../SmartProvider/types.js';
+import type { SmartProviderOptions } from '../SmartProvider/types.js';
 
 export type ProviderBuilderFn<P> = (
   rpcUrls: ChainMetadata['rpcUrls'],
   network: number | string,
-  retryOverride?: ProviderRetryOptions,
+  retryOverride?: SmartProviderOptions,
 ) => P;
 
 export type TypedProviderBuilderFn = ProviderBuilderFn<TypedProvider>;