+++ typescript/rebalancer-sim/CHANGELOG.md
@@ -1,5 +1,42 @@
 # @hyperlane-xyz/rebalancer-sim
 
+## 0.2.20
+
+### Patch Changes
+
+- Updated dependencies [b864cca]
+- Updated dependencies [1f918d0]
+- Updated dependencies [78199f4]
+  - @hyperlane-xyz/provider-sdk@5.1.0
+  - @hyperlane-xyz/sdk@33.0.2
+  - @hyperlane-xyz/utils@33.0.2
+  - @hyperlane-xyz/rebalancer@27.2.14
+  - @hyperlane-xyz/core@11.3.1
+
+## 0.2.19
+
+### Patch Changes
+
+- Updated dependencies [a2081df]
+- Updated dependencies [4c91737]
+- Updated dependencies [959c42a]
+  - @hyperlane-xyz/sdk@33.0.1
+  - @hyperlane-xyz/rebalancer@27.2.13
+  - @hyperlane-xyz/utils@33.0.1
+  - @hyperlane-xyz/core@11.3.1
+  - @hyperlane-xyz/provider-sdk@5.0.3
+
+## 0.2.18
+
+### Patch Changes
+
+- Updated dependencies [dc8e560]
+  - @hyperlane-xyz/sdk@33.0.0
+  - @hyperlane-xyz/rebalancer@27.2.12
+  - @hyperlane-xyz/utils@33.0.0
+  - @hyperlane-xyz/core@11.3.1
+  - @hyperlane-xyz/provider-sdk@5.0.2
+
 ## 0.2.17
 
 ### Patch Changes